requirements.txt
typing_extensions==4.10.0
whitenoise==6.6.0
python-dotenv==1.0.1
load_dotenv==0.1.0
django_heroku==0.3.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
django-cors-headers==4.3.1
